requirements.txt
Django==5.1.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
cryptography==42.0.5
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
stripe==11.0.0
instaloader==4.10.3
google-api-python-client==2.108.0
Pillow==10.2.0
